feat: create error module
