+++ app/controllers/application_controller.rb
@@ -6,10 +6,7 @@ class ApplicationController < ActionController::Base
   class Forbidden < ActionController::ActionControllerError; end
   class IpAddressRejected < ActionController::ActionControllerError; end
 
-  rescue_from StandardError, with: :rescue500
-  rescue_from Forbidden, with: :rescue403
-  rescue_from IpAddressRejected, with: :rescue403
-  rescue_from ActiveRecord::RecordNotFound, with: :rescue404
+  include ErrorHandlers
 
   private
 
@@ -19,18 +16,5 @@ class ApplicationController < ActionController::Base
     else
       'customer'
     end
-  end
-
-  def rescue403(e)
-    @exception = e
-    render 'errors/forbidden', status: 403
-  end
-
-  def rescue404(e)
-    render 'errors/not_found', status: 404
-  end
-
-  def rescue500(e)
-    render 'errors/internal_server_error', status: 500
   end
 end
